# test: use tiny random models to speed up tests