--- a/tests/evaluation/test_cmmd.py
+++ b/tests/evaluation/test_cmmd.py
@@ -13,7 +13,7 @@ from pruna.evaluation.task import Task
 @pytest.mark.parametrize(
     "model_fixture, device, clip_model",
     [
-        pytest.param("sd_tiny_random", "cuda", "openai/clip-vit-large-patch14-336", marks=pytest.mark.cuda),
+        pytest.param("ddpm-cifar10", "cuda", "openai/clip-vit-large-patch14-336", marks=pytest.mark.cuda),
         pytest.param("flux_tiny_random", "cpu", "openai/clip-vit-large-patch14-336", marks=pytest.mark.cpu),
     ],
     indirect=["model_fixture"],