--- a/tests/evaluation/test_energy_metrics.py
+++ b/tests/evaluation/test_energy_metrics.py
@@ -9,8 +9,8 @@ from pruna.evaluation.metrics.metric_energy import EnergyMetric
 @pytest.mark.parametrize(
     "model_fixture, device",
     [
-        pytest.param("stable_diffusion_v1_4", "cuda", marks=pytest.mark.cuda),
-        pytest.param("resnet_18", "cpu", marks=pytest.mark.cpu),
+        pytest.param("shufflenet", "cuda", marks=pytest.mark.cuda),
+        pytest.param("sd_tiny_random", "cpu", marks=pytest.mark.cpu),
     ],
     indirect=["model_fixture"],
 )