--- a/tests/evaluation/test_model_architecture_metrics.py
+++ b/tests/evaluation/test_model_architecture_metrics.py
@@ -10,8 +10,8 @@ from pruna.evaluation.metrics.metric_model_architecture import ModelArchitecture
 @pytest.mark.parametrize(
     "model_fixture, device",
     [
-        pytest.param("noref_resnet_18", "cpu", marks=pytest.mark.cpu),
-        pytest.param("stable_diffusion_v1_4", "cuda", marks=pytest.mark.cuda),
+        pytest.param("shufflenet", "cpu", marks=pytest.mark.cpu),
+        pytest.param("sd_tiny_random", "cuda", marks=pytest.mark.cuda),
     ],
     indirect=["model_fixture"],
 )