… tests/algorithms/test_combinations.py