## Description
Hugging Face offers “tiny-random” model versions for many popular models. This PR integrates these tiny random models into our test suite to speed up several test suites. On my g6 EC2 instance, the changes in this PR reduced the algorithm tests from **430 s → 215 s**, the combination tests from **322 s → 194 s**, and the evaluation tests from **570 s → 141 s**.

Moreover, because the random tiny models are much smaller, the initial download and setup for our tests is also significantly faster.

Since many of our tests (e.g., algorithm integration and combination tests) do not require models to generate meaningful outputs, using random tiny models is appropriate.

## Type of Change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

## How Has This Been Tested?
I ran all affected tests multiple times and confirmed they pass consistently with the tiny random models.

## Checklist
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes

## Additional Notes
This PR only modifies existing tests. Adopting tiny random models may also enable us to expand test coverage.