@johnrachwan123 2025-05-10
thanks a lot for the PR! This is very exciting! I was wondering if you know why the algorithm tests take 50% less time and not more ?