@nifleisch 2025-05-10
> thanks a lot for the PR! This is very exciting! I was wondering if you know why the algorithm tests take 50% less time and not more ?

To be honest, I was also surprised we didn’t see bigger speedups with these changes. Loading and smashing tiny random models in a script runs way more than twice as fast compared to doing the same with the original model. One factor that could have artificially improved the times for the old tests is that I ran the tests multiple times before measuring. Many of the larger models used in the test may have already been cached in RAM, so loading them was faster. We should investigate this further.